web_control/docker/requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
pyserial==3.5
python-engineio==4.9.0
python-socketio==5.11.0
numpy==1.24.3
